requirements.txt
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
docutils==0.16
html-to-csv==0.1.3
idna==2.9
kiwisolver==1.1.0
lxml==4.5.0
matplotlib==3.2.0
mysql-connector==2.2.9
mysqlclient==1.4.6
numpy==1.18.1
pandas==1.0.1
PyMySQL==0.9.3
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
schedule==0.6.0
six==1.14.0
soupsieve==2.0
SQLAlchemy==1.3.13
statistics==1.0.3.5
urllib3==1.25.8
